refactor: Implement lazy loading for CoursServerLayer component
