+++ src/app/cours/add/[slug]/page.tsx
@@ -1,38 +1,16 @@
-import AddCoursOrSequenceView from "@/features/cours-sequence/presentation/views/AddCoursView";
-import { compCatUsecases } from "@/features/comp-cat/application/usecases/comp-cat-usecases";
-import { authUseCases } from "@/features/auth/application/usecases/auth-usecases";
-import { isLeft } from "fp-ts/lib/Either";
-import { redirect } from "next/navigation";
-import { Competence } from "@/features/comp-cat/domain/entities/schemas";
 import NotFound from "@/app/not-found";
+import { Suspense } from "react";
+import CoursAddServerLayer from "./CoursAddServerLayer";
+import LoadingSkeleton from "@/core/components/common/LoadingSkeleton";
 
 export default async function Page({ params }: { params: { slug: string } }) {
   if (!params.slug) {
     return <NotFound />;
   }
 
-  const authUser = await authUseCases.getUserAuth();
-  if (isLeft(authUser)) {
-    redirect("/login");
-  }
-
-  const eitherCompetences = await compCatUsecases.getCompetences({
-    userId: authUser.right.userId,
-  });
-
-  let competences: Competence[] = [];
-  if (!isLeft(eitherCompetences)) {
-    competences = eitherCompetences.right;
-  } else {
-    competences = [];
-  }
   return (
-    <AddCoursOrSequenceView
-      competences={competences}
-      authUser={authUser.right}
-      sequenceId={params.slug}
-      type="cours"
-      title="Add Cours"
-    />
+    <Suspense fallback={<LoadingSkeleton />}>
+      <CoursAddServerLayer slug={params.slug} />
+    </Suspense>
   );
 }
